Create Superstore
This SQL query shows a superstore database that consists of the item name, identifier, category, price amount and rating. The 1st SQL query shows if we select all items from the database and ordered from the highest amount of items to the lowest amount of items. The 2nd SQL query shows pulling the max number of items in their categories from the database but grouping them by their name and ordering from highest to lowest.